Add clue numbering
This involved going back to nesting each cell's <input> element
inside of a <div> element. I don't love that; the grid seems somehow
cleaner as a collection of simple <input> elements. But I wanted to
add the clue numbers as ::after psuedo-elements. Since those are
technically nested inside the element they're attached to, and since
<input> elements cannot contain other elements...

I would rather sacrifice the "cleanliness" of an input-only grid,
than sacrifice the "elegance" of clue numbers as ::afters.