Possible fix for mobile rendering bug
We have an input, inside a div.cell, inside a div#palette.

With the div#palette's height set explicitly to n*cell-size, and the
div.cell's height implicit (as one of n*n divs inside the pallete,
which has display:grid, with an n*n layout, it should have height
cell-size), and the input's height at 100%, SOMEHOW on android chrome
we were getting the input with height n*cell-size.

This generally caused the entered text to be rendered either behind
another cell, or below the entire grid. Only on mobile chromium
though (Brave and Chrome). Desktop chromium, desktop firefox, and
mobile firefox were all handling it just fine.

Perhaps setting the height explicitly on the div.cell will fix this.